add default sorting; pagination; cleanup react components
